requirements.txt
apache-airflow-providers-google
apache-airflow-providers-apache-spark
apache-airflow==2.10.2
pyspark
